# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pokemon Anil API",
    "version": "2.0.1",
    "description": "API propria do Pokemon Anil/PT-BR. Dados locais Azul 4.0.6 (interna 4.0.3). Sem consultas externas nos catalogos de jogo."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/calculator": {
      "post": {
        "summary": "POST /calculator",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
    "/calculator/history": {
      "get": {
        "summary": "GET /calculator/history",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "atacante",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "defensor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      },
      "delete": {
        "summary": "DELETE /calculator/history",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "atacante",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "defensor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
    "/calculator/history/{id}": {
      "get": {
        "summary": "GET /calculator/history/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "writeKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}